PSREDEV-1829: Process all commit messages
When searching for the canary skip directive in commit messages, process all messages from the Pull Request. This allows the directive to be placed in any commit rather than in just the last one and reduce the chances that the canary deployment will be triggered accidentally where not intended.

For metadata, take the first line of the head commit message, which is effectively the subject line of the head commit.

Don't trim the head commit subject for metadata since S3 allows arbitrary Unicode characters in metadata values.
  Reference: https://docs.aws.amazon.com/AmazonS3/latest/userguide/UsingMetadata.html#UserMetadata

If the head commit message is not available in the GitHub Action (e.g. the trigger event is not push), get the last commit message from the git log.

Update the metadata descriptions in the README.