PSREDEV-1829: Add version input and deprecate version-number
Version can be an arbitrary string, not just a number, and is likely to contain some letters.

Add a new input called 'version' and add a deprecation message to 'version-number' to make this change backwards compatible.

Use 'version-number' as a fallback for 'version' for now.

Update input descriptions in the README.