PSREDEV-1910: Add a workflow to test the upload action
Add a test workflow:
  - Pull the demo SAM app repo
  - Build demo sam-app2
  - Use the trimmed branch name as the S3 prefix when uploading the artifact
  - Run the upload action to upload the build demo sam-app2 template
  - Run the test script

Output the packaged template path from the upload action and pass that in to the test script to ensure the correct template is used for testing.

Add a symlink to the action file to prevent IDE warnings.
The symlink is placed under the same directory name as the one the action repo is checked out into in the test workflow.
Since both the demo SAM app and the upload action repos need to be checked out in the test workflow, they need to be placed under named directories to prevent the GHA runner from overwriting the working directory and erasing the .aws-sam directory and previously checked out files.
This means the test workflow uses the ./upload-action directory in the runner to locate the upload action. When this is specified this way, the IDE shows warnings when viewing the test workflow since it can't locate an action under that directory locally. The symlink ensures the IDE can find the workflow and validate it locally.